Rework BookControllerTest

+++ app/src/main/java/com/book/library/book/AnyBookDto.java
@@ -1,3 +1,3 @@
 package com.book.library.book;
 
-public record AnyBookDto(long id, String name, String author, boolean borrowed) implements BookDto {}
+public record AnyBookDto(int id, String name, String author, boolean borrowed) implements BookDto {}
