+++ app/src/main/java/com/book/library/reader/ReaderRepository.java
@@ -2,4 +2,4 @@ package com.book.library.reader;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReaderRepository extends JpaRepository<Reader, Long> {}
+public interface ReaderRepository extends JpaRepository<Reader, Integer> {}
